Extract devtools-client-adapter
