+++ package.json
@@ -42,6 +42,7 @@
     "devtools-modules": "^0.0.9",
     "devtools-network-request": "^0.0.9",
     "devtools-sham-modules": "^0.0.9",
+    "devtools-client-adapters": "0.0.1",
     "documentation": "^4.0.0-beta11",
     "eslint": "^3.10.0",
     "expect.js": "^0.3.1",
