+++ packages/devtools-client-adapters/src/chrome.js
@@ -1,8 +1,8 @@
 /* eslint-disable */
 
 const { connect } = require("chrome-remote-debugging-protocol");
-const defer = require("../utils/defer");
-const { Tab } = require("../tcomb-types");
+const defer = require("./utils/defer");
+const { Tab } = require("./tcomb-types");
 const { isEnabled, getValue } = require("devtools-config");
 const networkRequest = require("devtools-network-request");
 const { setupCommands, clientCommands } = require("./chrome/commands");
