+++ packages/devtools-client-adapters/src/firefox/events.js
