+++ packages/devtools-client-adapters/src/utils/defer.js
@@ -1,0 +1,12 @@
+module.exports = function defer() {
+  let resolve, reject;
+  let promise = new Promise(function() {
+    resolve = arguments[0];
+    reject = arguments[1];
+  });
+  return {
+    resolve: resolve,
+    reject: reject,
+    promise: promise
+  };
+};
