+++ packages/devtools-local-toolbox/package.json
@@ -38,11 +38,11 @@
     "babel-register": "^6.18.0",
     "body-parser": "^1.15.0",
     "check-node-version": "^1.1.2",
-    "chrome-remote-debugging-protocol": "0.0.11",
     "classnames": "^2.2.5",
     "co": "=4.6.0",
     "css-loader": "^0.25.0",
     "devtools-config": "^0.0.9",
+    "devtools-client-adapters": "^0.0.1",
     "devtools-modules": "^0.0.9",
     "devtools-network-request": "^0.0.9",
     "devtools-sham-modules": "^0.0.9",
